Class Name Changes + MobBehaviour + NetworkOrchestrator
Change each class name so that if it deals with Photon Bolt is be called {name}Orchestrator. In addition to this, those classes must derive from NetworkOrchestrator for easy MobController Access and for clarity. MobBehaviour was added as a base class for child classes to be able to have an easy access to MobController.