Improve MobAnimationController.cs + Client-side Attack bug Fix
MobAnimationController is now reponsible for triggering animation events and looking up currently playing events. This takes over the role of AttackingComponent, which no longer needs to look at Animator raw. Finally, MobAnimationController is now included by default in MobBehaviour.
This improved class made it possible to fix a bug cleanly when the client attacked. Previously, the client didn't stop in place when attacking when connected to a server, but by directly checking if the attack animation is playing, the issue is fixed