Found the bug in Photon Bolt
Apparently, switching DLL modes can easily be broken in Photon bolt if both debug and release DLLs exist in Assets/Photon/PhotonBolt/assemblies. Here's normal behaviour:

If you're on the debug DLL, ONLY bolt.dll.release will exist in the directory. Conversely, if you're on the release DLL, ONLY bolt.dll.debug will exist in the directory.

HOWEVER, if the dll you're switching to doesn't exist, you get an error when switching DLLs. IF BOTH DLLs exist at the SAME TIME, no compilation will occur and you'll also be stuck on the same DLL.

TO FIX THIS, simply remove the DLL the CURRENT version is on.