Add DataComponent.cs
With this class, all references to a Mob's data can be traced to a single place. In addition to this, the class prevents any permanent changes to the data file by making a copy to it.

In addition to this change, PlayerMovementOrchestrator now inherits from MovementOrchestrator, as it somehow wasn't before.