Remove LocalScaleFlip from Mob and Player Prefabs
To make prefabs simpler, remove LocalScaleFlip from all prefabs. 

To do this, MobAnimationController now takes in a reference to SpriteRenderer, which has a built-in variable that flips the animation without messing with the scale of the entity. With this, all of the functionality in LocalScaleFlip can be accomplished in the main class.