Fix fatal sprite ordering errors
Usually, all sprites are supposed to be sorted according to their pivot. However, with trees, it was found that sorting was actually in the sprite's center instead of the pivot, so that option was changed.

In addition to this, the health bar of mobs was not in the top sorting layer, so it was set to Default at level 100.