Fix fatal Tiled Map map loading
The y-coordinates of Tiled are FLIPPED, so that negative is above the origin and positive is below. This makes sense in most games where a bare-bones graphics library is used; They all have a inverted y-axis. In Unity, however, the y-axis is not flipped, so the position of all mobs imported are flipped vertically

To fix this, simply put a negative sign on the MapEntityLoader.cs' entities to revert the y-flip