Fix Client bug where you attack when you type in chatbox
Previously, you attacked every single time you press space in the client, even if you were focused on the chatbox. This was because PlayerAttackOrchestrator.cs independently listened for input ONLY only on the client-side, making it unaffected by the lock in PlayerMovementOrchestrator.cs

To fix this, simply make PlayerAttachOrchestrator.cs not listen to input at all and only attack if it receives a command to do so. This indirectly couples it to PlayerMovementOrchestrator.cs, saving the day