Change API of AttackingComponent.cs
To increase cohesiveness, attacking component was changed of the following:

Add isAttacking bool to decrease mobAnimator coupling of checking the attack animation

Change Attack() to FrameAttack() to increase clarity on the function's purpose.

Change TriggerAttackEvent() to Attack() so that it's function is clearer. This now returns a bool that says whether or not the attack animation was able to be played if it wasn't already.