Rename ChatPlayerMovementLock.cs to ChatboxMovementLock.cs
There's no need to have "Player" in the name of this script as it's a UI script, meaning that there is an underlying assumption that only the player can see the chatbox.

By renaming it to ChatboxMovementLock.cs, it makes it more clear that the script it in the chatbox.