Add Common\Map, Map Bounds, and finite map
Added a script that automatically creates collision boxes (prefab called MapBound) around the map so that entities cannot escape.

To make this happen, you need the map width and height. Unfortunately, on infinite maps, you cannot view the size of the map. To get around this, simply uncheck the "Infinite" checkbox in Tiled to make map size visible.