Add MapLoader.cs for dynamic map loading
This was added so that maps can very easily be swapped with other maps, allowing rooms to use different maps without having to create a scene for each of them.

To make this happen however, the map must first be loaded into Unity, then saved as a prefab in Resources\Map. This is because the Asset SuperTiled2Unity is actually not a prefab, but instructions for a prefab. Due to this, ther must exist a separate asset of this map as a fully loaded unity prefab.