Fix spawning bug on death
PlayerDeathComponent.cs teleported the player to (0,0) if they died. However, these coordinates are on the topleft of the map, making it very impractical as an actual spawn zone.

To fix this, a SpawnRegistry.cs was created to provide a static reference to known spawn locations, allowing PlayerDeathComponent.cs to simply call it. Spawn locations are loaded in by MapEntityLoader.cs server side.

In addition to this, tight coupling between ServerCallbacks.cs and MapLoader.cs was completely eradicated with this new class.