Replace Current Collision System with Rigidbody2D
We want entities to collide with one another, which is just not possible with our current collision system. So, add a RigidBody2D to all our mobs and integrate it into MobController.cs. 

A manual Physics Updater was also added as prior experiments showed that it is needed in order for physics to be synced across machines