Adjust New Collision System so that entities sync properly
This is big; Listen up

Clients lagged behind the server. On closer inspection, this was due to the client's position not being updated the moment mobController was updated, leading to kickback.

The solution to this problem was to make Physics able to be manually updated, as implemented in the commit prior. While this completely solved the issue for the player, mobs and other players still had this issue only on the client side.

On closer inspection, this other issue was caused by noncontrolled entities not having their velocity synced with the server. As velocity is completely not needed to sync, the solution was to simply set their velocities to 0 in their movementOrchestrators if they are not being controlled.